update bindep for ubuntu 22.04
The python 3.10 tox job runs on ubuntu 22.04 in the A cycle ubuntu 22.04
will also become the default for other jobs. python-dev is nolonger packaged
on ubuntu 22.04 so the tox job fails.

This change sync the python deps to match those use in nova which
already works across all tested distros.

Change-Id: I7fa25fea03f0c9c645a1eb10b22f058ca548d2fb