Avoid rbac defaults conflict in functional tests
RBAC defaults for enforce_scope and enforce_new_defaults for
placement service is False. Nova is setting them for Nova service
to True in https://review.opendev.org/c/openstack/nova/+/866218

Placement tests fixture are used in Nova functional tests and
setting the rbac defaults causing placement to run with defaults
of what Nova is setting. To avoid this conflict, in placement
tests fixture, we need to set enforce_scope and enforce_new_defaults
to the same value it is for placement service.

Needed-By: https://review.opendev.org/c/openstack/nova/+/866218
Change-Id: I7e94d107962efdd7779733375327f5c8248c7087