[CI] Remove placement-nova-tox-functional-py37 job
Patch I1d6a2986fcb0435cfabdd104d202b65329909d2b removed job definition
of nova-tox-functional-py37 job from nova, and now that is missing for
placement-nova-tox-functional-py37 job. This patch removes the job to
unblock the gate.

This patch also bundles: Cap setuptools <71.0.0

py38 jobs (on ubuntu-focal) started to fail due to recent virtualenv
release (20.26.4) on Yoga (which bundles setuptools), because we have
'packaging==21.3' in this branch that is not compatible with newer
setuptools [1].

setuptools is bundled in virtualenv, so it has to be capped via the
virtualenv package. tox also needed to be capped (<4) as gate uses
tox 3.28.0, but with capping virtualenv we pull in latest tox as well,
which would cause other errors.

Change-Id: I2b70f34fa5387bb13caabdc8b1a1133271cb40d3