Add a global limit on the number of allocation candidates
When a root provider has multiple similar child RPs, i.e. we have a wide
but symmetric tree, and the allocation
candidate request asks for resources in multiple request groups the
number of possible allocation candidates grows rapidly.

* 1 root, 8 child RPs with 1 unit of resource each
  a_c requests 6 groups with 1 unit of resource each
  => 8*7*6*5*4*3=20160 possible candidates

* 1 root, 8 child RPs with 6 unit of resources each
  a_c requests 6 groups with 6 unit of resources each
  => 8^6=262144 possible candidates

Placement generates these candidates fully before applying the limit
parameter provided in the query to be able do a random sampling if
[placement]randomize_allocation_candidates is True.

Placement takes excessive time and memory to generates these amount of
allocation candidates.

To avoid request timeout or out of memory events this patch introduces a
new [placement]max_allocation_candidates config option that is applied
not after but *during* the candidate generation loop. So this new option
acts as a global safety limit.

The new config option is defaulted to -1, meaning no limit, to keep the
legacy behavior, but we suggest to tune that config in environments with
wide symmetric trees.

Placement generates the candidates with a depth first approach,
gathering all candidates for a root before moving to the next root. If
the new global limit is applied and hit then it is possible that the
returned candidates are from a limited set of root providers. This issue
will be handled by a subsequent patch allowing to select from
depth-first or breadth-first (round-robin) candidate generation
strategy.

Related-Bug: #2070257
Change-Id: I5475c66fa35958b3dc72c3bab7a7edbdf6c89ce6
(cherry picked from commit 9351241f91f80cb7dc45c4c0daab929b5dfc5c4b)