Adapt tests to new messages from  jsonschema 4.23.0
In the new jsonschema version couple of error messages are changed. As
the tests should work with both old and the new jsonschema version the
effected tests are relaxed to accept both error message.

See https://review.opendev.org/c/openstack/requirements/+/925059 for the
version bump patch where the tests are failing

Change-Id: I21e5f5f72a46560f25d254e569b8d5cd15126fda