improve test logging and replace psycopg2 with psycopg2-binary
This change replaces psycopg2 with psycopg2-binary removing
the need to have postgresql header files to compile the module
when running tests locally. nova did this years ago, so this
is just making placement use the binary wheel as well.

Additionally, this fixes how server-side logs are reported in
tox. Effectively we now use the OS_LOG_CAPTURE env var
to control if logs are printed. By default all logs to
stdout and stderr are captured and not printed to keep existing
behavior. You can now define OS_DEBUG=1 and OS_LOG_CAPTURE=0 to
change the  default log level to debug (from info) and allow
the logs to be printed when tests fail.

Change-Id: I576bdc102f10c6e50a5420590ac64f50dee124c0